ci: Add a workflow that updates casks at 4AM every day
We don't want to add this logic to the app repository to keep concerns separate and GitHub does not provide a way to trigger workflows automatically when a workflow from another repository succeeds so we went for a daily cron job.

At 4AM so the app should be built successfully if we create a nightly midnight job.